+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-
+- FIX Push batchSize in not respected [#3994](https://github.com/pubkey/rxdb/issues/3994)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->